+++ src/index.ts
@@ -1,5 +1,12 @@
 export { DatabaseClient, ResultType, Statement, query, sql } from './query'
-export { table, column, nullable, hasDefault, primaryKey } from './table'
+export {
+  table,
+  column,
+  nullable,
+  hasDefault,
+  primaryKey,
+  columnMapping,
+} from './table'
 export { date, boolean, integer, string, json } from './table/columns'
 export { Table, TableType, TableTypeWithDefaults, Column } from './table/types'