+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### master
 
+- add `selectAs` to rename selected columns
 - rename `selectAs` to `selectAsJson`
 - add `primaryKey` column marker to create correct group-by statements for `selectAsJsonAgg`
 - add `whereSql` and the `sql` tagged template function